fix sorting of datetimes for 'matches'
If given '--sortdt', 'scan' will sort lines of stream based on first
datetime (chronologically) on each given line, with the resulting stream
being used as input to other commands.
For 'matches', stream is sorted as before, after the sort the order of
datetimes, datetimes-as-strings, and datetime-positions, and if these
are not None, matches uses them instead of scanning the stream again.